Skip non-ASCII test on Windows
This test file causes failures to build as shutil does not
handle non-ASCII files well in Windows Python2.
On first run the test updates the file to use the unicode
filename, on subsequent runs we assume the renamed file as
this was the original case.

Skip non-ASCII test on Windows

This test file causes failures to build as shutil does not
handle non-ASCII files well in Windows Python2.
On first run the test updates the file to use the unicode
filename, on subsequent runs we assume the renamed file as
this was the original case.